requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.9
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.70.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-health-checking==1.80.0
grpcio-status==1.71.2
grpcio-tools==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
magic-filter==1.0.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
playwright==1.58.0
plotly==6.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.1
Pygments==2.20.0
pyparsing==3.3.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
squarify==0.4.4
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
validators==0.35.0
vnai==2.4.0
vnstock==3.5.0
vnstock_ezchart==0.0.3
watchdog==6.0.0
weaviate-client==4.20.4
websockets==16.0
wordcloud==1.9.6
yarl==1.23.0
zipp==3.23.0
